requirements.txt
Django
djangorestframework
django-cors-headers
drf-yasg
python-dotenv
psycopg2
dj-database-url
gunicorn
whitenoise
djangorestframework-simplejwt